function_calling_MCP/mcp-master/pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "googlesearch-python>=1.3.0",
    "mcp>=1.9.1",
    "openai>=1.78.1",
    "python-dotenv[cli]>=1.1.0",
    "spotipy>=2.25.1",
]
